+++ src/main/java/com/example/security/repository/mapper/OrderMapper.java
@@ -4,6 +4,7 @@ import com.example.security.model.Order;
 import com.example.security.model.Status;
 import org.springframework.jdbc.core.RowMapper;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -16,7 +17,7 @@ public class OrderMapper implements RowMapper<Order> {
         order.setAddressShipping(rs.getString("address_shipping"));
         order.setStatus(Status.valueOf(rs.getString("order_status")));
         order.setOrderDate(rs.getDate("order_date").toLocalDate());
-        order.setTotalPrice(rs.getDouble("total_price"));
+        order.setTotalPrice(BigDecimal.valueOf(rs.getDouble("total_price")));
         return order;
     }
 }
